ruff format

+++ utils/dynamodb_accessor.py
@@ -5,14 +5,13 @@ from boto3.dynamodb.conditions import Key
 
 
 class DynamoDBAccessor:
-
     def __init__(self, table_name, aws_access_key_id=None, aws_secret_access_key=None):
         if aws_access_key_id and aws_secret_access_key:
             self.dynamodb = boto3.resource(
                 'dynamodb',
                 aws_access_key_id=aws_access_key_id,
                 aws_secret_access_key=aws_secret_access_key,
-                region_name=os.environ['REGION']
+                region_name=os.environ['REGION'],
             )
         else:
             self.dynamodb = boto3.resource('dynamodb')
@@ -20,29 +19,22 @@ class DynamoDBAccessor:
 
     def get_item(self, tenant_id, data_id):
         try:
-            response = self.table.get_item(
-                Key={
-                    'tenant_id': tenant_id,
-                    'data_id': data_id
-                }
-            )
+            response = self.table.get_item(Key={'tenant_id': tenant_id, 'data_id': data_id})
             return response.get('Item')
         except Exception as e:
-            print(f"Error reading from DynamoDB: {e}")
+            print(f'Error reading from DynamoDB: {e}')
             return None
 
     def batch_get_item(self, tenant_id, data_ids):
         try:
             request_items = {
-                self.table.table_name: {
-                    'Keys': [{'tenant_id': tenant_id, 'data_id': data_id} for data_id in data_ids]
-                }
+                self.table.table_name: {'Keys': [{'tenant_id': tenant_id, 'data_id': data_id} for data_id in data_ids]}
             }
             response = self.dynamodb.batch_get_item(RequestItems=request_items)
             return response['Responses'][self.table.table_name]
 
         except Exception as e:
-            print(f"Error reading from DynamoDB: {e}")
+            print(f'Error reading from DynamoDB: {e}')
             return None
 
     def filter_by_field(self, tenant_id, field_name, field_value):
@@ -56,7 +48,7 @@ class DynamoDBAccessor:
             return response['Items']
 
         except Exception as e:
-            print(f"Error reading from DynamoDB: {e}")
+            print(f'Error reading from DynamoDB: {e}')
             return None
 
     def put_item(self, tenant_id, data_id, data, save_nested=True, metadata=None):
@@ -75,7 +67,7 @@ class DynamoDBAccessor:
 
             return True
         except Exception as e:
-            print(f"Error writing to DynamoDB: {e}")
+            print(f'Error writing to DynamoDB: {e}')
             return False
 
     def batch_put_item(self, tenant_id, data_ids: list, data: list, save_nested=True, metadata=None):
@@ -92,12 +84,10 @@ class DynamoDBAccessor:
                     if metadata:
                         item |= {'metadata': metadata}
 
-                    batch.put_item(
-                        Item=item
-                    )
+                    batch.put_item(Item=item)
             return True
         except Exception as e:
-            print(f"Error writing to DynamoDB: {e}")
+            print(f'Error writing to DynamoDB: {e}')
             return False
 
     def update_item(self, tenant_id, data_id, attribute_updates, metadata=None):
@@ -107,16 +97,13 @@ class DynamoDBAccessor:
 
             updated_expression, expression_attribute_values = get_update_params(attribute_updates)
             res = self.table.update_item(
-                Key={
-                    'tenant_id': tenant_id,
-                    'data_id': data_id
-                },
+                Key={'tenant_id': tenant_id, 'data_id': data_id},
                 UpdateExpression=updated_expression,
-                ExpressionAttributeValues=expression_attribute_values
+                ExpressionAttributeValues=expression_attribute_values,
             )
             return res['ResponseMetadata']['HTTPStatusCode'] == 200
         except Exception as e:
-            print(f"Error updating item in DynamoDB: {e}")
+            print(f'Error updating item in DynamoDB: {e}')
             return False
 
     def batch_update_item(self, tenant_id, data_ids: list, attribute_updates: list, metadata=None):
@@ -125,87 +112,77 @@ class DynamoDBAccessor:
                 attribute_update |= {'metadata': metadata}
 
         updated_expressions, expression_attribute_values = zip(
-            *[get_update_params(attribute_update) for attribute_update in attribute_updates])
+            *[get_update_params(attribute_update) for attribute_update in attribute_updates]
+        )
 
         updates = [
             {
                 'Update': {
                     'TableName': self.table.table_name,
-                    'Key': {
-                        'tenant_id': tenant_id,
-                        'data_id': data_id
-                    },
+                    'Key': {'tenant_id': tenant_id, 'data_id': data_id},
                     'UpdateExpression': updated_expression,
-                    'ExpressionAttributeValues': expression_attribute_value
+                    'ExpressionAttributeValues': expression_attribute_value,
                 }
             }
-            for data_id, updated_expression, expression_attribute_value in
-            zip(data_ids, updated_expressions, expression_attribute_values)
+            for data_id, updated_expression, expression_attribute_value in zip(
+                data_ids, updated_expressions, expression_attribute_values
+            )
         ]
         try:
-            response = self.dynamodb.meta.client.transact_write_items(
-                TransactItems=updates
-            )
+            response = self.dynamodb.meta.client.transact_write_items(TransactItems=updates)
             return response['ResponseMetadata']['HTTPStatusCode'] == 200
         except Exception as e:
-            print(f"Error updating item in DynamoDB: {e}")
+            print(f'Error updating item in DynamoDB: {e}')
             return False
 
     def delete_item(self, tenant_id, data_id):
         try:
-            response = self.table.delete_item(
-                Key={
-                    'tenant_id': tenant_id,
-                    'data_id': data_id
-                }
-            )
+            response = self.table.delete_item(Key={'tenant_id': tenant_id, 'data_id': data_id})
             return response
         except Exception as e:
-            print(f"Error deleting item from DynamoDB: {e}")
+            print(f'Error deleting item from DynamoDB: {e}')
             return None
 
     def batch_delete_item(self, tenant_id, data_ids: list):
         try:
             delete_items = {
-                self.table.table_name: [{'DeleteRequest': {'Key': {'tenant_id': tenant_id, 'data_id': data_id}}} for
-                                        data_id in data_ids]}
+                self.table.table_name: [
+                    {'DeleteRequest': {'Key': {'tenant_id': tenant_id, 'data_id': data_id}}} for data_id in data_ids
+                ]
+            }
             response = self.dynamodb.batch_write_item(RequestItems=delete_items)
             return response['ResponseMetadata']['HTTPStatusCode'] == 200
         except Exception as e:
-            print(f"Error deleting item from DynamoDB: {e}")
+            print(f'Error deleting item from DynamoDB: {e}')
             return None
 
     def list_items_by_tenant(self, tenant_id):
         try:
-            response = self.table.query(
-                KeyConditionExpression=boto3.dynamodb.conditions.Key('tenant_id').eq(tenant_id)
-            )
+            response = self.table.query(KeyConditionExpression=boto3.dynamodb.conditions.Key('tenant_id').eq(tenant_id))
             data_array = [item['data'] for item in response.get('Items', [])]
             return data_array
         except Exception as e:
-            print(f"Error querying DynamoDB: {e}")
+            print(f'Error querying DynamoDB: {e}')
             return []
 
     def list_data_ids_by_tenant(self, tenant_id):
         try:
             response = self.table.query(
                 KeyConditionExpression=boto3.dynamodb.conditions.Key('tenant_id').eq(tenant_id),
-                ProjectionExpression='data_id'
+                ProjectionExpression='data_id',
             )
             return [item['data_id'] for item in response.get('Items', []) if 'data_id' in item]
         except Exception as e:
-            print(f"Error querying DynamoDB: {e}")
+            print(f'Error querying DynamoDB: {e}')
             return []
 
     def get_all_items_by_tenant(self, tenant_id):
         try:
-            response = self.table.query(
-                KeyConditionExpression=boto3.dynamodb.conditions.Key('tenant_id').eq(tenant_id)
-            )
+            response = self.table.query(KeyConditionExpression=boto3.dynamodb.conditions.Key('tenant_id').eq(tenant_id))
             items_dict = {item['data_id']: item['data'] for item in response.get('Items', [])}
             return items_dict
         except Exception as e:
-            print(f"Error querying DynamoDB: {e}")
+            print(f'Error querying DynamoDB: {e}')
             return {}
 
 
@@ -219,11 +196,11 @@ def get_update_params(body):
     Returns:
         update expression, dict of values.
     """
-    update_expression = ["set "]
+    update_expression = ['set ']
     update_values = dict()
 
     for key, val in body.items():
-        update_expression.append(f" {key} = :{key},")
-        update_values[f":{key}"] = val
+        update_expression.append(f' {key} = :{key},')
+        update_values[f':{key}'] = val
 
-    return "".join(update_expression)[:-1], update_values
+    return ''.join(update_expression)[:-1], update_values
